hotel_service/hotels/serializers.py
from rest_framework import serializers
from hotel_service.hotels.models import Hotel


class HotelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Hotel
        fields = '__all__'
