pyproject.toml
[tool.poetry]
name = "hotelaggregatorservice"
version = "0.1.0"
description = ""
authors = ["Nikolaj Gordienko <nicolas.gordienko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
django-test = "^0.4030"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
